AST Validation to `common`
- For use in both `tile` and `api`.
